requirements.txt
PyQt5>=5.15.0
pandas>=1.5.0
requests>=2.25.0
beautifulsoup4>=4.9.0
pillow>=9.0.0